Expose publisher through DidChangeDecorator
Instead of installing a publisher on Store, we require the model to
conform to DidChangeProtocol, which is anything with a `didChange`
publisher property.

We offer a concrete implentation of DidChangeProtocol,
DidChangeDecorator, which can be used to decorate (wrap) one of your
models, usually the top-level model. This decorator class publishes all
actions passed to the `update` method directly after mutating the model
via the underlying update method.

DidChangeDecorator also exposes all properties of the wrapped model via
dynamic property access, so the interface from the outside remains the
same.

If the model conforms to this protocol, then the store exposes the
publisher, giving you the old API.

Why?

Subcomponents may want to subscribe to didChange. However, to expose a
didChange publisher on ViewStore using the old approach where store owns
the publisher would require two tagging functions — one to translate
from parent to child action, and another to translate from child to
parent. This feels like a bridge too far. OTOH, if we expose didChange
via the model update function, then the correctly scoped action will be
passed to the corresponding child model by virtue of the update function
being the thing that updates the model. We essentially get "scoped"
actions "for free" because no matter how the update is produced, it has
to go through the update function.

This approach means that exposing a publisher is opt-in. You have to
wrap your model in the decorator. That seems like an OK tradeoff.