Expose publisher through DidUpdateDecorator (#9)
This PR contains changes motivated by an initial pass at integrating with the web-based block editor.

- Publishers can be exposed at every level of the model tree (see design)
- Protocols are renamed to `Arboreal___` to avoid name collisions with ObservableStore

## Design

Instead of exposing a publisher on Store, we expose it on the model, through DidUpdateProtocol.

We offer a concrete implementation of DidUpdateProtocol through `DidUpdateDecorator<Model: ModelProtocol>`, which can be used to decorate (wrap) your models. This is usually the top-level model, but can be used at any level where you wish to publish actions for updates. This decorator class publishes all actions passed to the `update` method directly after mutating the model via the underlying update method.

If the model conforms to this protocol, then the store exposes the publisher, giving you the old API.

DidUpdateDecorator also exposes all properties of the wrapped model via dynamic property access, so the interface from the outside remains the same.

## Why?

Subcomponents using ViewStore sometimes want to subscribe to didUpdate. However, exposing a didUpdatet publisher on ViewStore presents some challenges:

- Using the old approach where the store owns the publisher would require two tagging functions — one to translate from parent to child action, and another to translate from child to parent.
- It would be insufficient to intercept actions in the ViewStore's send method, since updates to the same underlying model may come from other ViewStores, or from the root store.

Requiring two tagging functions feels like a bridge too far. What we need is to publish actions from a point that will always be hit regardless of where the model update is coming from. That place is the model's `update` function, not the `send` method.

If we expose didUpdate via the model, and publish to it from the model's update function, then the correctly scoped action will be passed to the corresponding child model by virtue of that update function being the very thing that updates the model. We essentially get "scoped" actions "for free" because no matter how the update is produced, it has to go through the update function.

## Tradeoffs

This approach means that exposing a publisher is opt-in. You have to wrap your model in the decorator. This tradeoff seems acceptable, since the benefits of being able to publish actions at any model level are large, and publishers may not always be desired or required.